Move LiveblogRenderer.tsx to components folder
